`@ConfigurationProperties`로 타입-안전한 설정 관리 도입 (#84)
* refactor: health관련 패키지 네이밍 및 파일 네이밍 변경

* feat(config): @ConfigurationProperties로 타입-안전한 설정 관리 도입

기존의 @Value나 하드코딩 방식은 오타에 취약하고, 설정값 누락 시
런타임 오류를 유발할 수 있는 타입 불안전성 문제가 있었습니다.

이를 해결하기 위해 @ConfigurationProperties를 사용하는 `FastApiProperties`
클래스를 도입하여 FastAPI 연동 설정을 중앙화하고, 애플리케이션 시작 시점에
설정값의 타입과 유효성을 검증하도록 개선했습니다.

이를 통해 잠재적인 런타임 장애를 원천 차단하고, 코드의 안정성과
유지보수성을 크게 향상시켰습니다.

* refactor: Code Formatting
